added parallel evaluation for speedup; moved scoring to isles/; converted config options to settings.py to allow for functions
