+++ Dockerfile
@@ -16,9 +16,11 @@ COPY --chown=evaluator:evaluator requirements.txt /opt/evaluation/
 RUN python -m pip install --user -r requirements.txt
 
 
-COPY --chown=evaluator:evaluator ground-truth /opt/evaluation/ground-truth
-COPY --chown=evaluator:evaluator config.json /opt/evaluation/config.json
+COPY --chown=evaluator:evaluator testlabels /opt/evaluation/ground-truth
 
 COPY --chown=evaluator:evaluator evaluation.py /opt/evaluation/
+COPY --chown=evaluator:evaluator settings.py /opt/evaluation/
+ADD --chown=evaluator:evaluator isles/ /opt/evaluation/isles/
+ADD --chown=evaluator:evaluator sample_bids/ /opt/evaluation/sample_bids/
 
 ENTRYPOINT "python" "-m" "evaluation"
