+++ isles/__init__.py
@@ -1,0 +1,3 @@
+from . import scoring
+
+__all__ = ['scoring']
